requirements.txt
cmd2>=0.8.2
jira>=1.0.11
pyyaml>=3.12
prompter>=0.3.10
python-editor>=1.0.3
attrs>=17.4.0
prettytable>=0.7.2
iso8601>=0.1.12
six>=1.11.0
pykerberos>=1.2.1
python.dateutil>=2.6.1
requests>=2.18.4
requests_kerberos>=0.12.0
setuptools_scm>=1.15.7
undecorated>=0.3.0
pathlib2>=2.3.0
wcwidth>=0.1.7
cached_property
-e .
